requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
pydantic==2.11.7
python-dotenv==1.1.1
pymysql==1.1.1
cryptography==45.0.5
redis==6.2.0
kafka-python==2.2.15